test(tui): share the production app fixture
